apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
slowapi>=0.1.9
python-multipart>=0.0.6
sqlmodel
asyncpg
supabase
langchain
langchain-openai
langchain-core
openai
anthropic
google-generativeai
python-dotenv
duckduckgo-search
fpdf2
greenlet